Add test/projects to gitignore.
test/projects is retrieved from the Scratch server and should be downloaded when using the testsuite for the first time. Even though we want to ensure that a specific version of the scratch package is test, we don't want to mirror the scratch package.

In future it might be a good idea to host only one project in that folder which is used entirely by the provided testsuite.